Rename package from duke to georgie
