+++ src/main/java/georgie/DialogBox.java
@@ -1,4 +1,4 @@
-package duke;
+package georgie;
 
 import java.io.IOException;
 import java.util.Collections;
