+++ src/main/java/georgie/Georgie.java
@@ -1,4 +1,4 @@
-package duke;
+package georgie;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
