+++ src/main/java/georgie/Main.java
@@ -1,4 +1,4 @@
-package duke;
+package georgie;
 
 import java.io.IOException;
 import java.util.Timer;
