+++ src/main/java/georgie/Storage.java
@@ -1,4 +1,4 @@
-package duke;
+package georgie;
 
 import java.io.BufferedReader;
 import java.io.File;
