+++ src/main/java/georgie/Task.java
@@ -1,4 +1,4 @@
-package duke;
+package georgie;
 
 /**
  * Represents a generic task.
