+++ src/main/java/georgie/Ui.java
@@ -1,4 +1,4 @@
-package duke;
+package georgie;
 
 import java.util.Scanner;
