+++ src/test/java/duke/ToDoTest.java
@@ -1,5 +1,8 @@
 package duke;
 
+import georgie.GeorgieException;
+import georgie.TaskList;
+import georgie.ToDo;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
